extracted common methods from the Main and Prompt classes into a DynamicContent module.
